corrected the review_2 comments
