+++ data/src/main/java/ru/fefu/data/calendar/di/CalendarModule.kt
@@ -6,15 +6,11 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import ru.fefu.data.calendar.repositoty.CalendarEventsDataRepository
 import ru.fefu.data.calendar.repositoty.CalendarEventsDataRepositoryImpl
-import ru.fefu.data.calendar.sources.CalendarDataSource
-import ru.fefu.data.calendar.sources.CalendarDataSourceI
 
 @Module
 @InstallIn(SingletonComponent::class)
 interface CalendarModule {
 
-    @Binds
-    fun provideDataSource(dataSourceImpl: CalendarDataSource): CalendarDataSourceI
     @Binds
     fun provideCalendarEventsRep(rep: CalendarEventsDataRepositoryImpl): CalendarEventsDataRepository
 }
